backend/agrovera_microservices/payment_service/requirements.txt
Django>=5.1
djangorestframework
psycopg2-binary
django-cors-headers
